Move Value implementation into Internal
